Fix FloppyDisk redefinition build error
Agent-Logs-Url: https://github.com/iblowmymind/Darkstar/sessions/c059fe82-f63e-4b86-abaf-4774a3dfb2a9

+++ src/iop/floppy_controller.cpp
@@ -304,7 +304,7 @@ void FloppyController::FinishWriteTrack() {
         int sectorCount = _trackBuffer.size() / sectorSize;
         if (sectorCount > 10) sectorCount = 10; // Limit to reasonable number
         
-        _drive->Disk->FormatTrack(Format::FM500, _track, 0, sectorCount, sectorSize);
+        _drive->Disk->FormatTrack(static_cast<int>(Format::FM500), _track, 0, sectorCount, sectorSize);
         _drive->Disk->SetModified();
     }
 }
